requirements.txt
# Core ML libraries
tensorflow-cpu>=2.19.0
torch>=1.9.0
transformers>=4.28.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# NLP libraries
nltk>=3.8.1
spacy>=3.5.0
gensim>=4.3.0

# CV libraries
opencv-python>=4.7.0
mediapipe>=0.9.1

# Audio processing
librosa>=0.10.0
soundfile>=0.12.1

# Flask for API
flask>=2.2.3
flask-cors>=3.0.10

# Data conversion/serialization
pydantic>=1.10.7
protobuf>=4.22.1

# Resume parsing
python-docx>=0.8.11
pdfminer.six>=20221105
PyMuPDF>=1.21.1

# Misc
tqdm>=4.65.0
matplotlib>=3.7.1
